requirements.txt
python-telegram-bot[rate-limiter]==20.1
openai  
tiktoken>=0.3.0
PyYAML==6.0.1
python-dotenv==0.21.0
pymongo[srv]==3.12
apscheduler
numpy

